Updated all of the Create classses and created a new mixin class in .utils
