+++ organizer/models.py
@@ -46,3 +46,5 @@ class NewsLink(models.Model):
     def __Str__(self):
         return "{}:{}".format(self.startup, self.title)
 
+    def get_absolute_url(self):
+        return self.startup.get_absolute_url()
