+++ organizer/templates/organizer/startup_detail.html
@@ -24,9 +24,11 @@
     </di>
 
 <p>{{ startup.description|linebreaks}}</p>
-{% if startup.newslink_set.all %}
+
 <section>
     <h3>Recent News</h3>
+
+{% if startup.newslink_set.all %}
     <u1>
         {% for newslink in startup.newslink_set.all %}
         <li>
@@ -34,19 +36,33 @@
         </li>
         {% endfor %}
     </u1>
-</section>
 {% endif %}
-{% if startup.blog_posts.all%}
+    <p>
+        <a href="{% url 'organizer_newslink_create' %}">
+            <button class="btn btn-primary btn-sm" onclick = "">
+                    Add Article</button>
+        </a>
+    </p>
+</section>
 <section>
+{% if startup.blog_posts.all%}
+<p>
     <h3>Blog Post{{ startup.blog_post.all|pluralize }}</h3>
     <u1>
         {% for post in startup.blog_posts.all %}
             <li>
-            <a href="{{post.get_absolute_url}}"> {{ post.title|title }}</a>
+            <a href="{{post.get_absolute_url}}">{{ post.title|title }}</a>
             </li>
         {% endfor %}
     </u1>
-</section>
+</p>
 {% endif%}
+    <p>
+        <a href="{% url 'blog_post_create' %}">
+        <button class="btn btn-primary btn-sm" onclick = "">
+                    Create Blog Post</button>
+    </p>
+</section>
+
 </article>
 {% endblock %}
