+++ organizer/views.py
@@ -2,7 +2,8 @@ from django.shortcuts import get_object_or_404, render, redirect
 from django.shortcuts import get_object_or_404, render
 from django.views.generic import View   # needed to make class based views
 from .models import Tag, Startup
-from .forms import TagForm, StartupForm
+from .forms import TagForm, StartupForm, NewsLinkForm
+from .utils import ObjectCreateMixin
 
 
 def tag_list(request):
@@ -21,21 +22,6 @@ def tag_detail(request, slug):
     #context = {'tag': tag}
     #return HttpResponse(template.render(context))
     return render(request,'organizer/tag_detail.html',{'tag':tag})
-
-class TagCreate(View):
-    form_class = TagForm
-    template_name = 'organizer/tag_form.html'
-
-    def get(self,request):
-        return render(request, self.template_name, {'form':self.form_class()})
-
-    def post(self, request):
-        bound_form = self.form_class(request.POST)
-        if bound_form.is_valid():
-            new_tag = bound_form.save()
-            return redirect(new_tag)
-        else:
-            return render(request, self.template_name, {'form':bound_form})
 
 '''This function is now rendered uselass as we have created the class above it
 def tag_create(request):
@@ -56,25 +42,21 @@ def tag_create(request):
 def startup_list(request):
     return render(request,'organizer/startup_list.html',{'startup_list': Startup.objects.all()})
 
-
 def startup_detail(request, slug):
     startup = get_object_or_404(Startup, slug__iexact=slug)
     return render(request,'organizer/startup_detail.html', {'startup':startup})
 
-class StartupCreate(View):
+
+class TagCreate(ObjectCreateMixin, View):
+    form_class = TagForm
+    template_name = 'organizer/tag_form.html'
+
+
+class StartupCreate(ObjectCreateMixin, View):
     form_class= StartupForm
     template_name = 'organizer/startup_form.html'
 
-    def get(self, request):
-        return render(request, self.template_name, {'form':self.form_class()})
 
-    def post(self, request):
-        bound_form = self.form_class(request.POST)
-        if bound_form.is_valid():
-            new_startup = bound_form.save()
-            return redirect(new_startup)
-        else:
-            return render(request, self.template_name,{'form': bound_form})
-
-
-
+class NewsLinkCreate(ObjectCreateMixin, View):
+    form_class = NewsLinkForm
+    template_name='organizer/newslink_form.html'
